--- a/setup.py
+++ b/setup.py
@@ -151,6 +151,7 @@ tests_requirements = [
     "mypy",
     "wsgidav",
     "crc32c",
+    "gdrivefs @ git+https://github.com/intake/gdrivefs.git",
 ]
 
 setup(