--- a/tests/func/test_tree.py
+++ b/tests/func/test_tree.py
@@ -288,6 +288,7 @@ def test_tree_upload_fobj(dvc, tmp_dir, cloud):
         pytest.lazy_fixture("azure"),
         pytest.lazy_fixture("gs"),
         pytest.lazy_fixture("webdav"),
+        pytest.lazy_fixture("gdrive"),
     ],
 )
 def test_tree_ls(dvc, cloud):