--- a/tests/remotes/gdrive.py
+++ b/tests/remotes/gdrive.py
@@ -1,7 +1,6 @@
 # pylint:disable=abstract-method
 import os
 import uuid
-from functools import partialmethod
 
 import pytest
 from funcy import cached_property
@@ -44,31 +43,6 @@ class GDrive(Base, CloudURLInfo):
     def get_url():
         # NOTE: `get_url` should always return new random url
         return "gdrive://" + GDrive._get_storagepath()
-
-    @cached_property
-    def client(self):
-        from gdrivefs import GoogleDriveFileSystem
-
-        os.makedirs(CREDENTIALS_DIR, exist_ok=True)
-        with open(CREDENTIALS_FILE, "w") as stream:
-            stream.write(os.getenv(GDRIVE_TEST_CREDENTIALS_DATA))
-
-        return GoogleDriveFileSystem(token="cache")
-
-    def mkdir(self, mode=0o777, parents=False, exist_ok=False):
-        if not self.client.exists(self.path):
-            self.client.mkdir(self.path)
-
-    def write_bytes(self, contents):
-        with self.client.open(self.path, mode="wb") as stream:
-            stream.write(contents)
-
-    def _read(self, mode):
-        with self.client.open(self.path, mode=mode) as stream:
-            return stream.read()
-
-    read_text = partialmethod(_read, mode="r")
-    read_bytes = partialmethod(_read, mode="rb")
 
 
 @pytest.fixture