@isidentical 2021-02-10
Timings (on the request of @efiop) looks ~same (beside the normal deviation)

master:
```console
 $ time dvc push dataset1
21 files pushed                                                                                                                     
dvc push dataset1  1,98s user 0,28s system 10% cpu 22,415 total                                                                     

 $ time dvc status -c
Cache and remote 'origin' are in sync.                                                                                              
dvc status -c  0,82s user 0,09s system 34% cpu 2,639 total

 $ rm -rf dataset1 .dvc/cache

 $ time dvc pull dataset1
A       dataset1/                                                                                                                   
1 file added and 20 files fetched                                                                                                   
dvc pull dataset1  1,76s user 0,31s system 26% cpu 7,689 total                                                                      

 $ rm -rf dataset1 .dvc/cache dataset1.dvc

 $ time dvc gc -c -f -a

WARNING: This will remove all cache except items used in the workspace and all git branches of the current repo.
No unused 'local' cache to remove.                                                                                                  
dvc gc -c -f -a  1,37s user 0,10s system 6% cpu 22,645 total                                                                        

 $ time dvc gc -c -f -a                                                               

WARNING: This will remove all cache except items used in the workspace and all git branches of the current repo.
No unused 'local' cache to remove.                                                                                                  
No unused cache to remove from remote.                                                                                              
dvc gc -c -f -a  0,90s user 0,08s system 33% cpu 2,940 total
```

this PR:
```console
 $ time dvc push dataset1
21 files pushed                                                                                                                     
dvc push dataset1  2,70s user 0,40s system 12% cpu 25,019 total                                                                     

 $ time dvc status -c    
Cache and remote 'origin' are in sync.                                                                                              
dvc status -c  0,84s user 0,08s system 35% cpu 2,592 total

 $ rm -rf dataset1 .dvc/cache 

 $ time dvc pull dataset1
A       dataset1/                                                                                                                   
1 file added and 20 files fetched                                                                                                   
dvc pull dataset1  1,61s user 0,29s system 24% cpu 7,898 total                                                                      
(.venv38) (Python 3.8.5+) [  4:42ÖS ]  [ isidentical@desktop:~/temp_dvc3(master✗) ]
 $ rm -rf dataset1 .dvc/cache dataset1.dvc 

 $ time dvc gc -c -f -a
WARNING: This will remove all cache except items used in the workspace and all git branches of the current repo.
No unused 'local' cache to remove.                                                                                                  
dvc gc -c -f -a  1,42s user 0,12s system 6% cpu 22,898 total                                                                        

 $ time dvc gc -c -f -a
WARNING: This will remove all cache except items used in the workspace and all git branches of the current repo.
No unused 'local' cache to remove.                                                                                                  
No unused cache to remove from remote.                                                                                              
dvc gc -c -f -a  0,87s user 0,10s system 30% cpu 3,173 total
```