@efiop 2021-02-11
For the record: @isidentical created a very handy test fixture using gdrivefs in https://github.com/iterative/dvc/pull/5427/files , but it needs a pydata token to work. I'll try to find the original creds we are using in CI and will add it to the secrets so we could bring that fixture back (maybe in the followup PR, not a blocker).